requirements.txt
# pip install -r requirements.txt

Flask
Pillow 
opencv-python
Click==7.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.6